Rebuild search.
Search is now called by method_missing. This allows to search for arbitrary data bags. I also added a convenience API. If only a string is passed instead of a Hash, it is search for with default parameters.

Also fixed is the escaping which now uses CGI::escape. The original escape somehow got injected into my test environment...